massive refactoring effort again
